userId-added

+++ src/services/contacts.js
@@ -8,10 +8,11 @@ export const getAllContacts = async ({
   sortOrder = 'asc',
   sortBy = '_id',
   filter = {},
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
-  const contactsQuery = Contact.find();
+  const contactsQuery = Contact.find(userId);
   const { contactType, isFavourite } = parseFilterParams(filter);
   if (contactType !== null && contactType !== undefined) {
     contactsQuery.where('contactType').equals(contactType);
@@ -19,7 +20,7 @@ export const getAllContacts = async ({
   if (isFavourite !== null && isFavourite !== undefined) {
     contactsQuery.where('isFavourite').equals(isFavourite);
   }
-  const contactsCount = await Contact.countDocuments();
+  const contactsCount = await Contact.countDocuments(userId);
   const contacts = await contactsQuery
     .skip(skip)
     .limit(limit)
@@ -33,21 +34,25 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactById = async (id) => {
-  return await Contact.findById(id);
+export const getContactById = async (id, userId) => {
+  return await Contact.findById({ _id: id, userId });
 };
 
-export const createContact = async (payload) => {
-  const contact = await Contact.create(payload);
+export const createContact = async (payload, userId) => {
+  const contact = await Contact.create(...payload, userId);
   return contact;
 };
 
-export const upsertContact = async (id, payload) => {
-  const contact = await Contact.findByIdAndUpdate(id, payload);
+export const upsertContact = async (id, payload, userId) => {
+  const contact = await Contact.findByIdAndUpdate(
+    { _id: id, userId },
+    payload,
+    { new: true },
+  );
   return contact;
 };
 
-export const deleteContact = async (contactId) => {
-  const contact = await Contact.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+  const contact = await Contact.findOneAndDelete({ _id: contactId, userId });
   return contact;
 };
